+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Use more efficient regex without lookahead ([#443](https://github.com/cucumber/cucumber-expressions/pull/443))
 - [Java] Support a leading `+` sign in `{float}`, `{double}` and `{bigdecimal}` ([#441](https://github.com/cucumber/cucumber-expressions/pull/441))
 - [Go] Parse floating point numbers with scientific notation ([#443](https://github.com/cucumber/cucumber-expressions/pull/443))
+- [Dart] Add a Dart implementation of Cucumber Expressions ([#444](https://github.com/cucumber/cucumber-expressions/pull/444))
 
 ## [20.0.0] - 2026-06-11
 ### Changed